mid-project-347359563/services/customer-service/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
python-dotenv
pyjwt
cryptography